move inner classes
